services/scrapers/events/ingest/requirements.txt
requests
pydantic
requests-cache
google-cloud-storage